+++ models/stock_set.py
@@ -8,8 +8,8 @@ class StockSet(db.Model):
     url = db.Column(db.String(), nullable=True)
     name = db.Column(db.String())
     ticker = db.Column(db.String())
-    first_price = db.Column(db.Integer())
-    last_price = db.Column(db.Integer())
+    first_price = db.Column(db.Numeric())
+    last_price = db.Column(db.Numeric())
     first_date = db.Column(db.Date())
     last_date = db.Column(db.Date())
